Full test works. Cleaned up code
