+++ src/main/java/bio/terra/controller/RepositoryApiController.java
@@ -11,7 +11,6 @@ import bio.terra.model.StudyModel;
 import bio.terra.model.StudyRequestModel;
 import bio.terra.model.StudySummaryModel;
 import bio.terra.service.DatasetService;
-import bio.terra.service.IngestService;
 import bio.terra.service.JobService;
 import bio.terra.service.StudyService;
 import bio.terra.validation.DatasetRequestValidator;
@@ -45,7 +44,6 @@ public class RepositoryApiController implements RepositoryApi {
     private final DatasetRequestValidator datasetRequestValidator;
     private final DatasetService datasetService;
     private final IngestRequestValidator ingestRequestValidator;
-    private final IngestService ingestService;
 
     @Autowired
     public RepositoryApiController(
@@ -56,8 +54,7 @@ public class RepositoryApiController implements RepositoryApi {
             StudyService studyService,
             DatasetRequestValidator datasetRequestValidator,
             DatasetService datasetService,
-            IngestRequestValidator ingestRequestValidator,
-            IngestService ingestService
+            IngestRequestValidator ingestRequestValidator
     ) {
         this.objectMapper = objectMapper;
         this.request = request;
@@ -67,7 +64,6 @@ public class RepositoryApiController implements RepositoryApi {
         this.datasetRequestValidator = datasetRequestValidator;
         this.datasetService = datasetService;
         this.ingestRequestValidator = ingestRequestValidator;
-        this.ingestService = ingestService;
     }
 
     @InitBinder
@@ -123,7 +119,7 @@ public class RepositoryApiController implements RepositoryApi {
     @Override
     public ResponseEntity<JobModel> ingestStudy(@PathVariable("id") String id,
                                                 @Valid @RequestBody IngestRequestModel ingest) {
-        String jobId = ingestService.ingestStudy(id, ingest);
+        String jobId = studyService.ingestStudy(id, ingest);
         return jobService.retrieveJob(jobId);
     }
