+++ src/main/java/bio/terra/flight/study/ingest/StudyIngestFlight.java
@@ -17,9 +17,7 @@ public class StudyIngestFlight extends Flight {
         BigQueryPdao bigQueryPdao = (BigQueryPdao)appContext.getBean("bigQueryPdao");
 
         addStep(new IngestSetupStep(studyDao));
-//        addStep(new IngestCreateStagingTableStep(studyDao, bigQueryPdao));
         addStep(new IngestLoadTableStep(studyDao, bigQueryPdao));
-// new step to add the row id column here
         addStep(new IngestRowIdsStep(studyDao, bigQueryPdao));
         addStep(new IngestInsertIntoStudyTableStep(studyDao, bigQueryPdao));
         addStep(new IngestCleanupStep(bigQueryPdao));
